requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
redis==5.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1   
